[todomvc] Polymer branch analyzer doesn't wrap on-key-up with KeyEvent
New to all this (who isn't?) but I changed the todomvc so it continues to work, using KeyboardEvent instead of KeyEvent. Given the wholesale changes in the analyzer, I'm not sure if we want to continue wrapping on-key-up and its ilk there, or if we want to make other changes?